removed redundant config files
